+++ adapters/AdColony/AdColonyAdapter/OpenBidding/GADMAdapterAdColonyRtbInterstitialRenderer.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2018 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,9 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Facebook mediation network adapter version.
-static NSString *const kGADMAdapterFacebookVersion = @"5.3.2.0";
+#import <Foundation/Foundation.h>
+#import <GoogleMobileAds/GoogleMobileAds.h>
 
-static NSString *const kGADMAdapterFacebookOpenBiddingPubID = @"placement_id";
+@interface GADMAdapterAdColonyRtbInterstitialRenderer : NSObject
 
-static NSString *const kGADMAdapterFacebookPubID = @"pubid";
+/// Asks the receiver to render the ad configuration.
+- (void)renderInterstitialForAdConfig:(nonnull GADMediationInterstitialAdConfiguration *)adConfig
+                    completionHandler:
+                        (nonnull GADMediationInterstitialLoadCompletionHandler)handler;
+
+@end