+++ adapters/DuAd/DuAdAdapter/GADMediationAdapterDuAd.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Facebook mediation network adapter version.
-static NSString *const kGADMAdapterFacebookVersion = @"5.3.2.0";
+#import <Foundation/Foundation.h>
+@import GoogleMobileAds;
 
-static NSString *const kGADMAdapterFacebookOpenBiddingPubID = @"placement_id";
+@interface GADMediationAdapterDuAd : NSObject <GADMediationAdapter>
 
-static NSString *const kGADMAdapterFacebookPubID = @"pubid";
+@end